pyproject.toml
[tool.poetry]
name = "Eks Cdk Stack"
version = "0.1.0"
description = "AWS infrastructure stack written in AWS CDK for Python"
authors = ["Li Lin <guru.lin@gmail.com>"]
license = "MIT"

# this section contains all dependencies required to run unit and integration tests
# not just dependencies requied for lambda runtime
#
# notes on versions of some dependencies
#
# the python and boto3 versions are deteremined by AWs Lambda runtime 
# https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html
#
# version of pydantic is based on what's used by aws-lambda-powertools
# https://github.com/awslabs/aws-lambda-powertools-python/blob/develop/pyproject.toml

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "1.20.32"
aws-cdk-lib = "^2.38.1"
constructs = "10.*"
pytest = "^6.2.5"
pytest-mock = "^3.5.1"
pytest-cov = "^3.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
# formatting, linting
black = "^22.3.0"
flake8 = "^4.0.1"
flake8-black = "^0.3.3"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.10.0"
flake8-debugger = "^4.1.2"
flake8-fixme = "^1.1.1"
flake8-isort = "^4.1.1"
flake8-variables-names = "^0.0.5"
isort = "^5.10.1"
pre-commit = "^2.19.0"
mypy = "^0.961"
types-setuptools = "^57.4.8"
types-toml = "^0.10.3"
boto3-stubs = "1.20.32"

[tool.isort]
profile = "black"
line_length = 120
src_paths = ["runtime", "infrastructure", "tests"]

[tool.black]
line-length = 120

[tool.pytest.ini_options]
minversion = "6.0"
# addopts = "--cov=runtime --cov-report html --cov-report xml"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["runtime"]
omit = ["tests/*", "infrastructure/*"]

[tool.coverage.html]
directory = "test_report"
